// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-context-branch-option-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface PhraseContextBranchOptionDto {
  alternativeToggleRef: string | null;
  boundaryKind: string | null;
  displayText: string;
  exactTokenId: number | null;
  isAlternativeSelected: boolean;
  passesThroughCount: number;
  selectionRef: string;
  sideEndsHereCount: number;
}