// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-context-side-page-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseContextBranchOptionDto } from '../models/phrase-context-branch-option-dto';
export interface PhraseContextSidePageDto {
  nextCursor: string | null;
  options: Array<PhraseContextBranchOptionDto>;
  passesThroughCount: number;
  sideEndsHereCount: number;
  totalOptions: number;
}