// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-selected-path-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseExactTokenDto } from '../models/phrase-exact-token-dto';
import { PhraseSelectedPathStepDto } from '../models/phrase-selected-path-step-dto';
export interface PhraseSelectedPathDto {
  endsAtBoundary: boolean;
  selectionRef: string | null;
  steps: Array<PhraseSelectedPathStepDto>;
  tokens: Array<PhraseExactTokenDto>;
}